+++ private/bufpkg/bufconnect/interceptors.go
@@ -16,6 +16,9 @@ package bufconnect
 
 import (
 	"context"
+	"errors"
+	"fmt"
+	"net/http"
 
 	"github.com/bufbuild/buf/private/pkg/app/applog"
 	"github.com/bufbuild/connect-go"
@@ -42,18 +45,31 @@ func NewCLIWarningInterceptor(container applog.Container) connect.UnaryIntercept
 	interceptor := func(next connect.UnaryFunc) connect.UnaryFunc {
 		return func(ctx context.Context, req connect.AnyRequest) (connect.AnyResponse, error) {
 			resp, err := next(ctx, req)
-			if resp != nil && resp.Header() != nil {
-				encoded := resp.Header().Get(CLIWarningHeaderName)
-				if encoded != "" {
-					if warning, err := connect.DecodeBinaryHeader(encoded); err == nil && len(warning) > 0 {
-						container.Logger().Warn(string(warning))
-					}
+			if resp != nil {
+				logWarningFromHeader(container, resp.Header())
+			} else if err != nil {
+				if connectErr := new(connect.Error); errors.As(err, &connectErr) {
+					logWarningFromHeader(container, connectErr.Meta())
 				}
 			}
 			return resp, err
 		}
 	}
 	return interceptor
+}
+
+func logWarningFromHeader(container applog.Container, header http.Header) {
+	encoded := header.Get(CLIWarningHeaderName)
+	if encoded != "" {
+		warning, err := connect.DecodeBinaryHeader(encoded)
+		if err != nil {
+			container.Logger().Debug(fmt.Errorf("failed to decode warning header: %w", err).Error())
+			return
+		}
+		if len(warning) > 0 {
+			container.Logger().Warn(string(warning))
+		}
+	}
 }
 
 // TokenProvider finds the token for NewAuthorizationInterceptorProvider.