+++ CHANGELOG.md
@@ -4,6 +4,22 @@
 
 - No changes yet.
 
+## [v1.23.1] - 2023-06-30
+
+- Fix issue where `buf beta graph` would not print modules within a workspace that
+  had no dependencies or dependents.
+- Fix issue where `buf beta graph` would print warnings for missing dependencies
+  that were actually present.
+
+## [v1.23.0] - 2023-06-29
+
+- Add `buf beta graph` to print the dependency graph for a module in DOT format.
+- Various small bug fixes.
+
+## [v1.22.0] - 2023-06-23
+
+- Change default for `--origin` flag of `buf beta studio-agent` to `https://buf.build`
+
 ## [v1.21.0] - 2023-06-05
 
 - Fix issue where locally-produced images did not have module information if the corresponding
@@ -898,7 +914,10 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.21.0...HEAD
+[Unreleased]: https://github.com/bufbuild/buf/compare/v1.23.1...HEAD
+[v1.23.1]: https://github.com/bufbuild/buf/compare/v1.23.0...v1.23.1
+[v1.23.0]: https://github.com/bufbuild/buf/compare/v1.22.0...v1.23.0
+[v1.22.0]: https://github.com/bufbuild/buf/compare/v1.21.0...v1.22.0
 [v1.21.0]: https://github.com/bufbuild/buf/compare/v1.20.0...v1.21.0
 [v1.20.0]: https://github.com/bufbuild/buf/compare/v1.19.0...v1.20.0
 [v1.19.0]: https://github.com/bufbuild/buf/compare/v1.18.0...v1.19.0